Replace BackGestureDispatcher with NavigationEventInput for iOS Integration
